requirements.txt

pandas
numpy
scikit-learn
matplotlib
joblib
kafka-python
